Add the test-files directory.
This is a replacement for the "data" directory in Matt Shotwell's repository.  Instead of keeping the URLs in an R dataset, I extracted the URLs to a text file.  I then wrote a simple bash script to download each of the URLs in the file using curl.

My hope is that this will make it more likely for people, who may not be R programmers, to download the test files.

Note that many of the URLs are no longer valid and none of them use HTTPS.  This is how they were in the original repository.