Remove the Software Prototype section.
I had removed the prototype source code from this project because it was written in R and not useful to people who don't know R.  The prototype has served its purpose and now there are many production-ready SAS7BDAT readers implemented in several languages.  At this point, a discussion of the prototype and the advantages of implementing it in R is not worth keeping in a specification.