Rework the description of the page types.
My intention was only to do one of the ToDo items, which was state that a SAS7BDAT has at most one 'meta' page.

This also fixes inconsistencies, like stating that there are three page types and documenting five.

It removes some redundancy and moves the description of the 'amd' page type to the 'Page Type' section, which is also where 'comp' is described.